Move DRF object permission class to a common module
